`:extra_slices` config
This enables support for configuring slices that are not in the
application's file structure, but come externally - usually shipped in a
gem. It uses `extra_slices` configuration option, which is inspired from
`extra_applications` in Elixir.

The config is a hash to give user ability to define slice name
themselves, to avoid possible naming conflicts.